blinkos & bareblink compile and work for millis & pixels
