+++ AS7/blink/bareblinkOS/bareblinkOS/blinkos.h
@@ -9,6 +9,7 @@
 // Gets us uintx_t
 // Not sure why, but it MUST be in the header and not the cpp
 #include <stdint.h>
+#include "pixelcolor.h"
 
 #define IR_FACE_COUNT 6
 #define PIXEL_FACE_COUNT 6
@@ -25,7 +26,8 @@ typedef uint32_t millis_t;
 
 
 // I know this is ugly, but keeping them in a single byte lets us pass them by value
-// and also lets us OR them together.
+// and also lets us OR them together. Efficiency in the updater trumps since it
+// runs every millisecond.
 
 #define BUTTON_BITFLAG_PRESSED          0b00000001
 #define BUTTON_BITFLAG_LONGPRESSED      0b00000010
@@ -57,7 +59,7 @@ struct loopstate_in_t {
 
     ir_data_buffer_t ir_data_buffers[IR_FACE_COUNT];
 
-    buttonstate_t buttonstate;
+    buttonstate_t buttonstate;   // Note that the flags in here are cleared after every pass though the loop
 
     uint8_t woke_flag;           // Have we slept and woke?
 
@@ -68,37 +70,10 @@ struct loopstate_in_t {
 
 };
 
-// Each pixel has 32 brightness levels for each of the three colors (red,green,blue)
-// These brightness levels are normalized to be visually linear with 0=off and 31=max brightness
-
-union pixelColor_t {
-    
-    struct {
-        uint8_t changed:1;
-        uint8_t r:5;
-        uint8_t g:5;
-        uint8_t b:5;
-    };
-    
-    uint16_t as_uint16;
-    
-    pixelColor_t(uint8_t r_in , uint8_t g_in, uint8_t b_in );
-    
-};
-
-inline pixelColor_t::pixelColor_t(uint8_t r_in , uint8_t g_in, uint8_t b_in ) {
-    
-    r=r_in;
-    g=g_in;
-    b=b_in;
-    
-}
-
-
 struct loopstate_out_t {
 
-    pixelColor_t colors[PIXEL_FACE_COUNT];  // Current RGB colors of all faces
-                                            // Topmost bit set indicates it changed
+    pixelColor_t colors[PIXEL_FACE_COUNT];  // Current RGB colors of all faces. 5 bits each.
+                                        // Topmost bit set indicates it changed
 
     uint8_t futureexapansion[20];  // Save some space for future expansion
 
@@ -109,5 +84,5 @@ extern loopstate_in_t loopstate_in;
 
 // These are provided by blinklib
 
-void setupEntry();
+//void setupEntry();
 void loopEntry( loopstate_in_t const *loopstate_in , loopstate_out_t *loopstate_out);
